requirements.txt
selenium==4.11.2
pandas==1.5.3
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.4.2